requirements.txt
python_irodsclient==1.1.1
enum34
six
humanize>=0.5
dnspython>=2.2.0
backports.functools-lru-cache>=1.6.4
typing_extensions==4.1.1
PyYaml
